fix: switch from hook to utility function to avoid vue dependency
