+++ packages/themes/src/index.ts
@@ -12,7 +12,7 @@ import solarizedTheme from './presets/solarized.css?inline'
 import baseVariables from './variables.css?inline'
 
 export { migrateThemeVariables } from './utilities/legacy'
-export { useObtrusiveScrollbars } from './hooks/useObtrusiveScrollbars'
+export { hasObtrusiveScrollbars } from './utilities/hasObtrusiveScrollbars'
 
 export const themeIds = [
   'alternate',
